pyproject.toml
[project]
name = "browser-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Ilya05228", email = "iliyaand05@gmail.com" }]
requires-python = ">=3.14"
dependencies = [
    "camoufox[geoip]>=0.4.11",
    "pyinstaller>=6.18.0",
    "pyside6>=6.10.2",
    "numpy>=1.24.0",           # Требуется для camoufox
    "playwright>=1.58.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
browser-automation = "browser_automation.main:main"
main = "browser_automation.main:main"
gui = "browser_automation.gui_modern:main"
server = "browser_automation.server:start_server"
test-api = "browser_automation.test_instagram_api:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
